pf/15: Port vm_dealloc finder to 15.5+
Sometime after iOS 15.0, two additional arguments were added to the
function called at the detected location, both set to zero, which means
that this call site doesn't call vm_dealloc anymore but a different,
related function.

The easiest way out is to find another vm_dealloc call site. The
callsite that I picked works on iOS 15.0 and 15.5+. In versions 15.1 -
15.4, direct calls to vm_deallocate apparently didn't exist.

Originally, I included three more instructions in the pattern, before
the bl, but these turned out not to exist in the iphone8,1 kernel, for
example. The instructions:

         0x5280002a,     /* mov w10, #1 */
         0x1ac92149,     /* lsl w9, w10, w9 */
         0x8b29c102,     /* add x2, x8, w9, sxtw */